requirements-dev.txt
flake8
coverage
mypy
twine
google-crc32c
-r requirements.txt
